Conflicts:
	robot/arm_hal/hal/frontCamera.cpp
	robot/robot.uvopt
	robot/supervisor/src/visionSystem.cpp

+++ include/anki/cozmo/robot/hal.h
@@ -347,7 +347,7 @@ namespace Anki
 
       // Starts camera frame synchronization (blocking call)
       void CameraGetFrame(u8* frame, CameraMode mode,
-          u16 exposure, bool enableLight);
+          f32 exposure, bool enableLight);
 
       // Get the number of lines received so far for the specified camera
       //u32 CameraGetReceivedLines(CameraID cameraID);
